Extract usePwaInstallPrompt from App.tsx
First in a series of low-risk hook extractions to chip away at the
1817-line App.tsx component. This one moves the
beforeinstallprompt capture, the installPrompt state, and the
prompt()/userChoice flow into src/hooks/usePwaInstallPrompt.ts.

The post-install side effect (re-requesting notification permission)
stays in App.tsx as an onInstalled callback so the hook stays free
of `user` and `fcmToken` knowledge.

53/53 tests pass. App.tsx loses ~15 lines.
